Column Visibility
renamed and reorganized files
persist column visibility
parse table cell values based on type
