+++ viewer/app/app.scss
@@ -47,8 +47,7 @@ $color-purple-dark  : #66057A;
 @import '../components/pagination/pagination';
 @import './modules/search/search';
 @import './modules/health/health';
-@import './modules/session/styles/session';
-@import './modules/session/styles/colheader';
+@import './modules/session/styles/session.list';
 @import './modules/session/styles/session.sticky';
 @import './modules/session/styles/session.graph';
 @import './modules/session/styles/session.actions';
