+++ viewer/app/modules/index.js
@@ -4,9 +4,8 @@
 require('./user/user.service');
 require('./user/has.permission');
 
-require('./session/components/colheader.component');
-require('./session/components/session.component');
 require('./session/services/session.service');
+require('./session/components/session.list.component');
 require('./session/components/session.map.component');
 require('./session/components/session.graph.component');
 require('./session/components/session.detail.component');
@@ -18,8 +17,6 @@ require('./session/components/session.scrub.pcap.component');
 require('./session/components/session.delete.component');
 require('./session/components/session.send.component');
 require('./session/components/session.info.component');
-require('./session/components/colvis.component');
-require('./session/components/table.header.component');
 
 require('./search/components/search.component');
 require('./search/components/expression.typeahead');
