--- viewer/app/modules/session/styles/colheader.scss
@@ -1,27 +1,0 @@
-.moloch-col-header {
-  padding     : 0 !important; // padding applied to child div instead
-  white-space :nowrap;
-  overflow    :visible;
-
-  div {
-    height  : 46px;
-    padding : 15px 5px 5px 5px;
-
-    &.active {
-      color: $color-purple-dark;
-      background-color: $color-gray-light;
-    }
-  }
-
-  .fa-sort { visibility: hidden; }
-
-  &:hover {
-    .fa-sort { visibility: visible; }
-  }
-
-}
-
-.col-vis-menu {
-  max-height: 300px;
-  overflow  : auto;
-}
