--- viewer/app/modules/session/templates/table.header.html
@@ -1,40 +1,0 @@
-<tr ng-if="tableState">
-  <th width="6%">
-    <colvis visible-headers="tableState.visibleHeaders"></colvis>
-  </th>
-  <th ng-repeat="header in headers" colheader class="moloch-col-header"
-    col-name="{{::header.friendlyName}}" col-id="{{::header.dbField}}"
-    order="tableState.order"></th>
-  <th width="20%">Info</th>
-  <!-- <th width="6%">
-    <colvis visible-headers="tableState.visibleHeaders"></colvis>
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Start" col-id="fp" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Stop" col-id="lp" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Src IP" col-id="a1" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Src Port" col-id="p1" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Dst IP" col-id="a2" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Dst Port" col-id="p2" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Packets" col-id="pa" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Bytes" col-id="by" order="tableState.order">
-  </th>
-  <th colheader class="moloch-col-header"
-    col-name="Node" col-id="no" order="tableState.order">
-  </th>
-  <th width="20%">Info</th> -->
-</tr>
